aoc-2020-12-14: switch to Vec when expanding addresses
Change the return type of expand() from HashSet to Vec. While a set is
better from an architectural point of view, the repeated merges of
collections make the faster Vec the better choice in this case. Besides,
the outer function floating_addresses() still returns a HashSet.

Performance gains: 89.7 ms -> 57.3 ms:

  Previous version:

  $ hyperfine target/release/aoc-2020-12-14
  Benchmark #1: target/release/aoc-2020-12-14
    Time (mean ± σ):      89.7 ms ±   4.4 ms    [User: 86.5 ms, System: 3.3 ms]
    Range (min … max):    85.9 ms … 108.2 ms    27 runs

  This version:

  $ hyperfine target/release/aoc-2020-12-14
  Benchmark #1: target/release/aoc-2020-12-14
    Time (mean ± σ):      57.3 ms ±   2.7 ms    [User: 55.2 ms, System: 2.1 ms]
    Range (min … max):    54.9 ms …  65.8 ms    46 runs

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>