AOC 2020 day 20
Not the best code I've written, but at least it works. Took me forever
to figure out that the puzzle pieces are made of transparent glass and
so can be flipped upside down and still match, and that the same is true
for the finished puzzle. The example solution is only one of the eight
permutations of valid solutions!

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>