Restructure project directories
To better support multiple years of AOC challenges, nest each year in a
directory of its own. Also drop the <month> part from the project names.

Example:

  name : aoc-2020-12-01  ->  name : aoc-2020-01
  dir  : aoc-2020-12-01  ->  dir  : 2020/01

Also, update the template accordingly, and fix a bug when constructing
the URL to the input of a challenge (strip leading zeroes in the <day>
part).

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>