AOC 2015 day 22
Puzzles with too many corner cases quickly turn to messy code, and I
never have the energy to go back and improve the code. This is the case
with today's puzzle.

The code searches a tree (branched by available spells) depth-first. A
generic solution would have searched breadth-first, but to keep the code
simple, the code relies on a maximum depth parameter to limit the size
of the tree; the value was found by trial and error.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>